backend/requirements.txt
Flask==2.0.3
PyJWT==2.1.0
requests==2.26.0
python-dotenv==0.19.2
SQLAlchemy==1.4.25
Flask-Mail==0.9.1
twilio==6.62.0